+++ acceptance/bundle_test.go
@@ -109,7 +109,11 @@ func TestAppleAppBundleParity(t *testing.T) {
 				if err != nil {
 					t.Fatal(err)
 				}
-				got = strings.ReplaceAll(got, filepath.Join(portable, "Contents/MacOS/hello"), "<executable>")
+				portablePath, err := filepath.EvalSymlinks(filepath.Join(portable, "Contents/MacOS/hello"))
+				if err != nil {
+					t.Fatal(err)
+				}
+				got = strings.ReplaceAll(got, portablePath, "<executable>")
 				want = strings.ReplaceAll(want, nativePath, "<executable>")
 				if got != want {
 					t.Fatalf("display %s\nGo:\n%s\nApple:\n%s", option, got, want)